# src: add fast api calls to push span strings