--- a/lib/_http_client.js
+++ b/lib/_http_client.js
@@ -481,10 +481,9 @@ function ClientRequest(input, options, cb) {
     api.context.with(api.trace.setSpan(api.context.active(), span), () => {
       fn.call(this);
       // At this point we know the request is actually being initiated
-      const u = `${protocol}//${this.getHeader('host')}${this.path}`;
       span._pushSpanDataString(kSpanHttpMethod, method);
       span._pushSpanDataString(kSpanHttpProtocolVersion, '1.1');
-      span._pushSpanDataString(kSpanHttpReqUrl, u);
+      span._pushSpanDataString3(kSpanHttpReqUrl, protocol, this.getHeader('host') || 'localhost', this.path);
       this.prependOnceListener('error', requestOnError);
       this.once('close', requestOnClose);
     }, this);