--- a/lib/_http_server.js
+++ b/lib/_http_server.js
@@ -1113,8 +1113,7 @@ function parserOnIncoming(server, socket, state, req, keepAlive) {
   let api;
   let span;
   if (generateSpan(kSpanHttpServer)) {
-    const protocol = req.connection.encrypted ? 'https' : 'http';
-    const u = `${protocol}://${req.headers.host || 'localhost'}${req.originalUrl || req.url}`;
+    const protocol = req.connection.encrypted ? 'https:' : 'http:';
     api = getApi();
     const tracer = api.trace.getTracer('http');
     const ctxt = extractSpanContextFromHttpHeaders(api.ROOT_CONTEXT,
@@ -1128,7 +1127,7 @@ function parserOnIncoming(server, socket, state, req, keepAlive) {
     req[nsolid_span_id_s] = span;
     span._pushSpanDataString(kSpanHttpMethod, req.method);
     span._pushSpanDataString(kSpanHttpProtocolVersion, req.httpVersion);
-    span._pushSpanDataString(kSpanHttpReqUrl, u);
+    span._pushSpanDataString3(kSpanHttpReqUrl, protocol, req.headers.host || 'localhost', req.originalUrl || req.url);
   }
 
   if (req.upgrade) {