--- a/lib/internal/otel/trace.js
+++ b/lib/internal/otel/trace.js
@@ -8,7 +8,6 @@ const {
 const binding = internalBinding('nsolid_api');
 const { getSpanId,
         getTraceId,
-        pushSpanDataString,
         nsolid_consts } = binding;
 
 const {
@@ -38,7 +37,7 @@ class SpanContext {
   }
 }
 
-function Span(internalId, ids_str, spanContext, startTime, type, kind, links) {
+function Span(internalId, parentSpanId, spanContext, startTime, type, kind, links) {
   this._spanContext = spanContext;
   this.ended = false;
   this.internalId = internalId;
@@ -56,7 +55,11 @@ function Span(internalId, ids_str, spanContext, startTime, type, kind, links) {
       this._pushSpanDataUint64(nsolid_consts.kSpanKind, this.kind);
     }
 
-    this._pushSpanDataString(nsolid_consts.kSpanOtelIds, ids_str);
+    this._pushSpanDataString(nsolid_consts.kSpanTraceId, spanContext.traceId);
+    this._pushSpanDataString(nsolid_consts.kSpanSpanId, spanContext.spanId);
+    if (parentSpanId) {
+      this._pushSpanDataString(nsolid_consts.kSpanParentSpanId, parentSpanId);
+    }
   }
 }
 
@@ -210,7 +213,15 @@ Span.prototype._pushSpanDataDouble = function(type, name) {
 
 Span.prototype._pushSpanDataString = function(type, name) {
   if (this._isSampled()) {
-    pushSpanDataString(this.internalId, type, name);
+    binding.pushSpanDataString(this.internalId, type, name);
+  }
+
+  return this;
+};
+
+Span.prototype._pushSpanDataString3 = function(type, val1, val2, val3) {
+  if (this._isSampled()) {
+    binding.pushSpanDataString3(this.internalId, type, val1, val2, val3);
   }
 
   return this;
@@ -240,7 +251,6 @@ class Tracer {
       }
     }
 
-    let ids_str;
     let traceState;
     let traceId;
 
@@ -248,12 +258,10 @@ class Tracer {
     if (!parentContext) {
       // New root span.
       traceId = getTraceId();
-      ids_str = `${traceId}:${spanId}`;
     } else {
       // New child span.
       traceId = parentContext.traceId;
       traceState = parentContext.traceState;
-      ids_str = `${traceId}:${spanId}:${parentContext.spanId}`;
     }
 
     const internalId = newInternalSpanId();
@@ -269,7 +277,7 @@ class Tracer {
 
     const startTime = options.startTime || now();
     const span = new Span(internalId,
-                          ids_str,
+                          parentContext?.spanId,
                           spanContext,
                           startTime,
                           options.type || nsolid_consts.kSpanCustom,