--- a/src/node_external_reference.h
+++ b/src/node_external_reference.h
@@ -94,6 +94,19 @@ using CFunctionBufferCopy =
                  uint32_t source_start,
                  uint32_t to_copy);
 
+using CFunctionPushSpanDataString =
+    void (*)(v8::Local<v8::Object> receiver,
+             uint32_t trace_id,
+             uint32_t type,
+             const v8::FastOneByteString& val);
+using CFunctionPushSpanDataString3 =
+    void (*)(v8::Local<v8::Object> receiver,
+             uint32_t trace_id,
+             uint32_t type,
+             const v8::FastOneByteString& val1,
+             const v8::FastOneByteString& val2,
+             const v8::FastOneByteString& val3);
+
 // This class manages the external references from the V8 heap
 // to the C++ addresses in Node.js.
 class ExternalReferenceRegistry {
@@ -125,6 +138,8 @@ class ExternalReferenceRegistry {
   V(CFunctionWithBool)                                                         \
   V(CFunctionBufferCopy)                                                       \
   V(CFunctionWriteString)                                                      \
+  V(CFunctionPushSpanDataString)                                               \
+  V(CFunctionPushSpanDataString3)                                              \
   V(const v8::CFunctionInfo*)                                                  \
   V(v8::FunctionCallback)                                                      \
   V(v8::AccessorNameGetterCallback)                                            \